Read logo dimensions through one place that understands SVG
getimagesize() only reads raster headers and returns false for an SVG, which
the shop logo is allowed to be. Five places destructured its result without
checking: the front controller's logo variables, the invoice and supply-order
PDF headers, the webservice header-image PUT, and the upload itself. Each of
them got two nulls, and on PHP 8.5 a "Cannot use bool as array" diagnostic
with them.

The SVG reading that was added for the upload is moved to
Core\Image\ImageDimensions, which always answers with a pair of integers, and
the other four call it too. It reads an SVG's own width and height when they
are absolute and falls back to its viewBox otherwise, so those callers now get
real dimensions rather than nothing.

The webservice call also destructured a type and an attribute string it never
used; they are gone with it.
