+++ classes/pdf/HTMLTemplateSupplyOrderForm.php
@@ -3,6 +3,7 @@
  * For the full copyright and license information, please view the
  * docs/licenses/LICENSE.txt file that was distributed with this source code.
  */
+use PrestaShop\PrestaShop\Core\Image\ImageDimensions;
 
 /**
  * @deprecated since 9.0 and will be removed in 10.0, stock is now managed by new logic
@@ -168,7 +169,7 @@ class HTMLTemplateSupplyOrderFormCore extends HTMLTemplate
         $width = $height = 0;
 
         if (!empty($path_logo)) {
-            list($width, $height) = getimagesize($path_logo);
+            [$width, $height] = ImageDimensions::of($path_logo);
         }
 
         $this->smarty->assign([
